working docker access form git user inside git-receiver container

+++ bootstrap.zip/setup-contaiers-configuration/image-files/git-receiver/launch-new-version.sh
@@ -23,8 +23,10 @@ function rebuild_app_image() {
 function stop_existing_app() {
   echo -e "\n* [BUILDER] stoping exisitng app containers"
   APP_CONTAINERS=$(docker ps | grep "app-image" | cut -d' ' -f1)
-  docker stop $APP_CONTAINERS
-  docker rm -v $APP_CONTAINERS
+  if [ -z "$APP_CONTAINERS" ]; then
+    docker stop $APP_CONTAINERS
+    docker rm -v $APP_CONTAINERS
+  fi
 }
